+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Quit grafterm with `Esc` key.
 - User defined datasources via flag and/or env var.
 - Alias flag to override dashboard datasource ID using user datasource IDs.
 - Fallback dashboard referenced datasources to user datasources.