+++ CHANGELOG.md
@@ -9,6 +9,10 @@
 - Alias flag to override dashboard datasource ID using user datasource IDs.
 - Fallback dashboard referenced datasources to user datasources.
 
+### Fixed
+
+- Gauges that had color thresholds not being show.
+
 ## [0.1.0] - 2019-05-13
 
 ### Added